# NEON performance improvements