I tested out @maltanar changes for additional NEON accumulators from https://github.com/maltanar/libpopcnt/commits/neon-accumulators and it results in a minor performance improvement with sizes over 1KB on my ARMv7a platform:

**master**
```
Iters: 10000000
Array size: 16.00 KB
Algorithm: NEON
Status: 100%
Seconds: 67.11
2.4 GB/s
```
**neon-accumulators**
```
Iters: 10000000
Array size: 16.00 KB
Algorithm: NEON
Status: 100%
Seconds: 62.65
2.6 GB/s
```
**master**
```
Iters: 10000000
Array size: 1.00 KB
Algorithm: NEON
Status: 100%
Seconds: 5.55
1.8 GB/s
```
**neon-accumulators**
```
Iters: 10000000
Array size: 1.00 KB
Algorithm: NEON
Status: 100%
Seconds: 5.37
1.9 GB/s
```
**neon-accumulators**
```
Iters: 10000000
Array size: 4.00 KB
Algorithm: NEON
Status: 100%
Seconds: 17.01
2.4 GB/s
```
**master**
```
Iters: 10000000
Array size: 4.00 KB
Algorithm: NEON
Status: 100%
Seconds: 17.87
2.3 GB/s
```
**master**
```
Iters: 10000000
Array size: 128 bytes
Algorithm: NEON
Status: 100%
Seconds: 1.97
0.7 GB/s
```
**neon-accumulators**
```
Iters: 10000000
Array size: 128 bytes
Algorithm: NEON
Status: 100%
Seconds: 2.12
0.6 GB/s
```