@kimwalisch 2017-11-07
Hi Kyle,

I have failed twice to install an AArch64 Linux on my raspberry pi 3 today, I have done it before but I have forgotten how...

I have a few questions:

Why do you want to improve the ARM NEON code?
Just for fun or are you actually using it in one of your projects?