@maltanar 2017-11-07
Hi Kyle, Kim. Just FYI, the reason why I didn't end up re-submitting that PR is (if I remember correctly) a bug in the implementation. It's been a while and I don't have the opportunity to go check it at once, but I think it was about the accumulators overflowing if the bit string contained all 1s or too many 1s or something like that.