@kimwalisch 2017-11-07
Hi Yaman,

Thanks for your report. Personally I am undecided whether your improvement is worth it, your new ARM NEON algorithm uses twice as much code for a minor improvement in some circumstances (i.e. array sizes > 4 KB). That's why I did not modify the ARM NEON code myself back in May 2017.

I am now waiting for @KOConchobhair's feedback...