@KOConchobhair 2017-11-08
I'm actually using this in one of my projects. Fast popcnt is critical to my algorithm so I will take any speed up I can get, especially on ARM. In my testing, it does not seem to have any bugs but I did not specifically test the all 1's case. I'll try to do that soon.