@kimwalisch 2017-11-08
> Fast popcnt is critical to my algorithm so I will take any speed up I can get, especially on ARM.

OK, that's a good argument I can understand.

I will continue to try to reiinstall Aarch64 Linux on my raspberry pi 3 and when done I will try to clean up and tweak the improved ARM NEON code. Concerning the bug that @maltanar mentioned we shoud try to test huge arrays > 4 Gigabytes using the new ARM NEON code. This test case is not covered by libpopcnt's test suite.