@kimwalisch 2017-11-08
I have now successfully installed Aarch64 Linux on my raspberry pi 3 and I am now able to test the new ARM NEON code. On my system the new ARM NEON code is also slightly faster (up to 15%).

The bug reported by @maltanar might be related to:

```C++
    /* each iteration can increment each 8-bit accumulator by at most 16
     * (2 popcounts of an 8-bit number). up to 15 iters is safe, since 15*16 = 240 < 255.
     *  sum and clear the accumulators every 15th iteration before they overflow.
     * since 15 = (2^4)-1, can use bitwise AND to calculate modulo.
     */
```

Here @maltanar assumes that the count accumulators are reset every 15 iterations which would prevent the byte counts to overflow 255. Note that the critical test case is an array where all bits are set to 1. My test suite only tests arrays with random data so my test suite is not testing this critical test case.

I think @maltanar assumption is wrong since his code does 16 (not 15) iterations before the counters are reset:

```C++
    if((i & 15L) == 15L) {
      sum = vpadalq_u32(sum, vpaddlq_u16(vpaddlq_u8(t0)));
      sum = vpadalq_u32(sum, vpaddlq_u16(vpaddlq_u8(t1)));
      sum = vpadalq_u32(sum, vpaddlq_u16(vpaddlq_u8(t2)));
      sum = vpadalq_u32(sum, vpaddlq_u16(vpaddlq_u8(t3)));
      t0 = vcombine_u8(vcreate_u8(0), vcreate_u8(0));
      t1 = vcombine_u8(vcreate_u8(0), vcreate_u8(0));
      t2 = vcombine_u8(vcreate_u8(0), vcreate_u8(0));
      t3 = vcombine_u8(vcreate_u8(0), vcreate_u8(0));
    }
``` 

The expression ```(i & 15L)``` may be 0, 1, 2, 3, ..., 15 which are 16 iterations (not 15) after which the counters are reset. Because of this issue @maltanar code would not return the correct result for an array filled with only 1 bits that is larger than 128 bytes * 16 = 2 kilobytes.

I have not actually tested any of my assumptions above but if they are correct a possible code fix would be to reset the counters every 8 iterations:

```C++
    if((i & 7) == 7) {
    ...
```