@kimwalisch 2017-11-08
I have now added the critical test case for an array filled with only 1 bits to [test/test1.cpp](https://github.com/kimwalisch/libpopcnt/blob/master/test/test1.cpp#L55) on the master branch. The test now correctly **fails** for the new ARM NEON code!

The good news is however that my assumption that we can fix the new ARM NEON code using ```(i & 7) == 7```is correct, using this change the modified ```test1.cpp``` passes.

I will play with the new ARM NEON code over the next days, try to polish and tweak it...