@kimwalisch 2017-11-09
I have now merged @maltanar's [faster ARM NEON algorithm](https://github.com/kimwalisch/libpopcnt/blob/master/libpopcnt.h#L698) with my [bug fix](https://github.com/kimwalisch/libpopcnt/blob/master/libpopcnt.h#L712). According to my benchmarks it runs up to 20% faster if the array size is smaller or equal to the CPU's L1 cache size (32 kilobytes).

Thanks @maltanar and @KOConchobhair.